Move zero-fill responsibility back to front end
I thought it’d be better to zero-fill empty target.investigated_as in the back end, but I no longer think so. But the code I rewrote to start to allow this turns out to be more efficient than I had, in that I have only one dictionary search instead of two. ENCD-3213